requirements.txt
geemap
earthengine-api
openrouteservice
pandas
openpyxl
folium
jinja2
pytest
pytest-mock
pyyaml
